+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 0.1.0 (unreleased)
 
+- Add support for same `l3_virtual_network` under multiple `anycast_gateways`
+- Add `fabric_zone` support and adding `l3_virtual_networks` under fabric zone
 - BREAKING CHANGE: Add support to update `authentication_template` settings globally and on fabric site level
 - Add `ASSIGN` state for assigning device to site without provisioning
 - Fix pnp issue when adding multiple devices